+++ methods/src/main/kotlin/methods/fractionroots/FractionRootsRules.kt
@@ -175,7 +175,7 @@ enum class FractionRootsRules(override val runner: Rule) : RunnerMethod {
             val term2 = oneOf(integer2, radical2)
             // a * root[x, 3] + b * root[y, 3]
             val signedTerm2 = optionalNegOf(term2)
-            val denominator = ConditionPattern(sumOf(term1, signedTerm2)) { _, match ->
+            val denominator = ConditionPattern(sumOf(term1, signedTerm2)) { _, match, _ ->
                 match.isBound(radical1) || match.isBound(radical2)
             }
 
@@ -276,7 +276,7 @@ enum class FractionRootsRules(override val runner: Rule) : RunnerMethod {
             val radical2 = withOptionalIntegerCoefficient(integerOrderRootOf(UnsignedIntegerPattern()), true)
             val term2 = oneOf(integer2, radical2)
 
-            val denominator = ConditionPattern(sumOf(term1, term2)) { _, match ->
+            val denominator = ConditionPattern(sumOf(term1, term2)) { _, match, _ ->
                 match.isBound(radical1) || match.isBound(radical2)
             }
 
@@ -314,7 +314,7 @@ enum class FractionRootsRules(override val runner: Rule) : RunnerMethod {
             val term2 = oneOf(integer2, radical2)
 
             val signedTerm2 = optionalNegOf(term2)
-            val denominator = ConditionPattern(sumOf(term1, signedTerm2)) { _, match ->
+            val denominator = ConditionPattern(sumOf(term1, signedTerm2)) { _, match, _ ->
                 match.isBound(radical1) || match.isBound(radical2)
             }